commit by bash

+++ user/fork.c
@@ -152,11 +152,11 @@ int make_shared(void *va) {
 			perm |= PTE_LIBRARY;
 			syscall_mem_map(0, addr, 0, addr, perm);
 		}
-		return (*vpt)[PTX(addr)] & (~0xfff);
+		return ((u_long)(*vpt))[PTX(addr)] & (~0xfff);
 	}
 	if ((r = syscall_mem_alloc(0, addr, (PTE_V | PTE_R | PTE_LIBRARY))) < 0)
 		return -1;
-	return (*vpt)[PTX(addr)] & (~0xfff);
+	return ((u_long)(*vpt))[PTX(addr)] & (~0xfff);
 }
 
 /*** exercise 4.9 4.15***/
